give controller with/without SES options
There are two ways to approach this. I picked the one where we pass a
filename into addVat(). If we're using SES then we use rollup to convert it
into a single string, then pass it into s.evalate(). If we're
"native" (outside SES) then we pass it to a dynamic require() call.

The other approach would be to pass a function into addVat(). The
native/non-SES controller would just use it directly, and the SES flavor
would have to stringify it. This seems more challenging because it forces
authors to write their code in a single wrapper function, without imports.
But we might flip back to that approach if other factors change in the
future.