fix tests, remove dispatch.subscribe
Previously, the kernel avoided using dispatch.subscribe (to tell a Vat that
the kernel is interested in the resolution of some promise for which that Vat
is the decider) until some other Vat used syscall.subscribe (to tell the
kernel that that Vat was interested in the resolution of some promise). This
worked ok when the kernel knew who was the decider, which was the case for
promises created as the answer to a delivery (to some export), and for
promises exported by a vat (e.g. when sending a local promise as an
argument).

But when the promise was created as the answer to a pipelined message (a
syscall.send to a promise, not to an export), there is not yet a decider. It
was too hard to keep track of when the promise got redirected to someone who
could make a decision.

So I removed dispatch.subscribe, and now all Vats that find themselves
holding a resolver will immediately use .then on the matching promise, so the
callback/errback will notify the kernel about its resolution. Vats still use
syscall.subscribe to indicate that they want to be notified, but now the
kernel itself is always notified.

The whole idea of doing lazy subscription is meant to reduce the number of
messages flying around when someone sends a message but ignores the
answer (i.e. they drop the answer promise before calling .then on it). This
would depend upon being able to tell when .then is called, which would either
require using non-Native Promises (so we could wrap .then), having a platform
trick to get notified when .then got called, and/or having weakrefs so we
could detect that the answer promise went away without anyone calling .then .
As we have none of these things, the complexity burden of retaining the
ability to use them just wasn't worth it.