installability hack, add require()-able main.js
I'm having problems installing the published tarball from other packages, and
it's somehow tied to the dependency on our (internal) '@agoric/evaluate'
package. As a mitigation, I'm commiting agoric-evaluate-0.0.1.tgz in the
source tree (which is generally a bad idea), and telling SwingSet that it
should get dependency from a tarball instead of a source directory. I'm also
adding it to bundledDependencies so downstream users won't need their own
copy of the tarball.

Also, I added a `main:` to `package.json` that `require()` can use, since
apparently `module:` isn't sufficient even when both the importer and the
importee are using ES6-module syntax.