finish updating names, get closer on txn sending
This correctly fetches the account/sequence number first, but only if the
ag-solo's address is already known to the chain (the x/auth module must have
the key in the accounts table, which happens if it was given tokens in the
genesis block, or if it receives tokens later via a balance transfer). This
means we'll need an outside-of-swingset mechanism to introduce the chain to
new clients, probably our controller node: before it tells the chain-side
swingset machine to add egress entries for the new client, it must tell the
chain to accept messages from the new client.

Something is still broken in the transaction delivery, I'm not sure what.

The remaining occurrences of "sscli" have been replaced with
"ag-cosmos-helper".