WIP: rewrite state load/store handling
Works with a local solo vat, but I haven't even tried to run the tests yet.

Solo machines should create an empty object, wrap it with
makeStorageInMemory(obj), and attach it to config.externalStorage . The
kernel will modify this object in place, and after each cycle you can
JSON.stringify it off to disk. The next time the machine starts up,
pre-populate this object with the contents from disk.

In a chain environment, set config.externalStorage to something with
get/set/has -type wrapper methods.