add Presence.toString(), remove the marker method
Presences are nominally empty objects: they're only special because they live
in a table that gets consulted by E() or x!foo(). But for debugging it's
handy to know that you've got a Presence and not some other object. We used
to do this by adding a dummy method (named `_importID_NNN`): we can't add
data properties or the presence becomes a record and can't be serialized
properly.

This replaces that marker method with a plain `toString()`, which also means
a console.log() shows something useful. Thanks to mfig for the improvement.