phase 1, only kernel changes
"phase 1" means:

* change the kernel objectID/promiseIDs formats from the old `{type,
  id, [vat]}` to o+14, ko4, p-5, etc
* add the kernel Object Table to track object owners, instead of recording
  the owning vat ID inside each c-list entry
* replace the kernel/vat c-lists to point at ObjectTable / PromiseTable
  entries, simplify/robustify implementation
* kernel/controller APIs (queueToExport,addImport,addExport) take new refs
* update tests to match

The following changes are part of #88 but will be implemented in a subsequent
phase:

* allocate kernel promises in response to vat-supplied p+NN references
* remove syscall.createPromise()
* allow vats (specifically the comms vat) to opt-in to pipelining

The new kernel/vat state vectors are still built out of plain JS objects.
After we replace the kernel serialization mechanism (and no longer require
fast JSON.stringify() of the whole state) we can switch to Maps.