tests: new framework to exercise various message patterns
including pipelining.

This will absorb the comms-integration tests, and will exercise the same
patterns on both local-only and comms-connected vats. I'm trying to make the
test format as concise as possible, so we feel no hesitation to add lots of
them.

Some of the non-comms patterns are showing pipeline-esque behavior anyways,
because of the relative ordering of promise resolution: if the initial
promise in a pipeline chain is resolved before the second message reaches the
front of the queue, the second message will be delivered right away instead
of being queued inside the promise, which results in the same delivery order
as real pipelining would provide.

To demonstrate non-pipelining in the local (non-comms) case, I had to add a
new pattern (a72) which explicitly delays that first resolution until all the
pipelined messages have made it through the kernel queue. That pattern shows
pipelining when run through comms vats, but not when run locally.