switch to new 'CapData' and 'Message' types: kernel changes only
Previously, each syscall/dispatch API (and lots of internal interfaces, like
the run-queue) used an inconsistent variety of ways to pass the two halves of
"data that can include capability references": the string of serialized data,
plus the array of slots (like 'o+4' and 'p-4').

This defines a type (to the extent that one can define types in Javascript)
named "CapData" to hold "capability-bearing data". Instances of this type
contain exactly `{ body, slots }`. All API signatures that used to pass two
arguments now pass a single CapData-typed argument.

It also defines a type named "Message", which contains `{ method, args,
result }`, where `method` is a string that names the method to be invoked,
`args` is a CapData whose top-level member is an array (of positional
arguments), and `result` is either a promise reference or `null`/`undefined`.

"Message" is not revealed to Vats, but the various in-kernel run-queues and
delivery APIs are now written in terms of Messages rather than distinct
method / args / result fields.

This depends upon a new version of `@agoric/marshal` which uses CapData in
serialize/unserialize, rather than separate string and slots arguments.

refs #28