Use (mostly numeric) vatID/deviceID everywhere within kernel
.. as the index for kernel state data structures, instead of user-provided
vatName/deviceName strings. This will make it easier to use key/value state
storage (#144), since the keys will be more uniform (and shorter, and not
vulnerable to people using the separator string in their vat name). They look
like "vNN" and "dNN", where NN is a positive integer.

This will also make adding dynamic Vats easier (#19), as we'll just increment
a counter instead of asking the user to invent a unique name.

closes #146