rewrite storage API, use host-supplied DB-like object
This removes c.getState(). Instead, the host should retain control over the
hostDB object it provides to the controller, so the host can choose when the
hostDB should commit a block's worth of changes.

The kernel's Keepers use a read-cache to minimize cross-Realm calls and
hostDB operations. This read-cache does not yet evict any entries, so a
future task is to build some basic LRU-like policy for it. But the largest
performance problem right now, the vat transcript, is specifically *not* kept
in the cache, so memory usage should be reduced somewhat even without proper
eviction.

refs #144

Extra thanks to @gamedevsam in #164 for a recommendation in kernelKeeper, to
replace for-loop starting points with named constants.