swingset storageWrapper: fix cache-heuristic bug
Chip noticed this during review: the cacheHeuristic() function (which decides
whether it is useful to cache a given key) wasn't behaving as I intended.
Transcript entries should not be cached, since they're effectively write-only
at runtime. But the counter that keeps track of which entry we're adding is
worth caching. It's pretty inconsequential (and the entire kernel-side cache
should be ripped out anyways), but it might as well behave as originally
intended.