autoswap-frontend: temporarily rename package to avoid conflict
There's an existing `packages/agoric-cli/template/ui` which declares itself
to be `autoswap-frontend`, and I want to add `packages/autoswap-frontend`
with the same name. I suspect the new one will supplant the old one, but I'm
not sure. So for now I'll rename the `agoric-cli` template to be named
`autoswap-frontend-2`, and we can resolve the conflict properly later.