chore: fix (still-broken) test of liveslots promise handling
The previous test was failing because of #823, but would have continued to
fail even after that got fixed because the targets were mixed up. This ought
to pass once #823 is done.

These tests will be improved further as part of #859, as they aren't yet
systematically checking all the cases we identified.