feat: Add a notifier facility for Zoe and contracts
Creates a new package for notifier.

The Notifier has 2 facets: notifier and updater. The updater methods for
updateState() and resolve. The notifier has getUpdatesSince().
updateState() would be used within a service to record state changes and
resolve to record a final state and close the stream. getUpdatesSince()
would be made available to clients to allow them to be notified of state
changes by waiting for promise resolution.

Includes usage documentation and tests.