chore: hush the cosmic-swingset/ag-solo warning about harden+-non-SES
When `@agoric/harden` is imported outside a SES environment, it logs a
warning that `harden()` cannot provide the level of protection that it's
documented to provide. The specific problem is that `harden()` only freezes
out to the "fringe" (which is initially filled with all the primordials like
`Object.prototype`), and if we aren't running under SES, these primordials
are not already frozen. So `harden()` outside SES doesn't protect you against
prototype poisoning.

We use `harden()` is lots of code, because it's the right thing to do, and
some of that code (like `@agoric/assert`) gets used both inside and outside
of SES.

There's a special way to tell `@agoric/harden` that "yes, I know I'm not
using SES, but that's ok, please don't print your warning". It is only
appropriate to send this signal from the top-most file of an application (the
initial `main.js` executable). Libraries should never send this signal,
because they are not responsible for setting up a SES environment (nor are
they usually aware of running inside one).

This patch sends this signal (by adding a falsy `harden` property to
`globalThis`) in the top-most executable for `ag-solo`. This hushes the
warnings caused by using `@agoric/assert` and others in the non-kernel-Realm
code used by ag-solo.

The kernel-Realm SwingSet code uses a different `globalThis`, so the warning
is not disabled there. But the warning isn't triggered there anyways, because
it really does use SES.

refs #971

Issue #971 will be fixed properly when we move to new-SES and make the
top-most file responsible for calling `lockdown()` to establish a global SES
environment. At this point all primordials should be frozen, and `harden()`
should be equally safe everywhere.