chore(transform-metering): start to remove old-SES code
The tests were changed to use `install-ses` to create a SES environment. One
test was partially disabled until I figure out how to test the new-SES
environment properly (it exercises evaluation of the transformed cde, and was
dependent upon getting an old-SES tamed global environment). My plan for this
test is to exercise just the source transformation, and have a separate test
which exercises evaluation.

This should not affect downstream usage.