chore(swing-store-lmdb): switch tests to install-ses
LMDB is somewhat incompatible with SES: it uses an `Error`-modifying trick to
locate the shared library that it needs to dynamically load into the process
at startup, and of course SES freezes `Error`. Until SES finds a way to
tolerate this, the workaround is to perform a dummy `import 'node-lmdb'`
before importing `@agoric/install-ses`. The test now exercises this, but
programs that use swingset (and thus swing-store-lmdb) must do it too.

This does not affect the API, so it should not be visible to downstream code,
however that code *will* see the change in the SwingSet semver number and
must update to match (by installing SES at startup).