fix(zoe)!: replace evaluate with import-bundle, new metering
Zoe's contract evaluator was rewritten to use `import-bundle` and the new
metering system. This requires a SES environment, so any downstream code
which uses zoe must be run in an application that does `import
'@agoric/install-ses'` (or `install-metering-and-ses`) at startup.

The API has changed slightly: the `install()` method now takes a single
argument named `bundle`, whereas previously it took a tuple of `(source,
moduleFormat)`. The `bundle` argument should be exactly the output of `await
bundleSource(path)`. The internal contents of the bundle are private to
`bundle-source` and `import-bundle`: future changes will be made in both
simultaneously. However `install()` retains backwards compatibility for the
same of external dapps, until we get them all rewritten.

The `getInstallation()` method now returns the entire bundle, not just the
source code component. When comparing bundles (to make sure the installation
points to the code you're expecting), you must do a deep comparison of the
bundle objects, rather than a shallow comparison of source code strings.

The `makeZoe()` constructor no longer needs to be given `require`, but old
dapps (in other repos) might still provide it. We arrange our endowments to
make sure a leftover `require` argument won't interfere with the one we
create locally, so `makeZoe` is backwards compatible with those dapps.

However, if metering is desired, `makeZoe` requires a second argument named
`vatPowers` which is normally obtained from the vat's `buildRootObject()`
arguments. Zoe will pull `transformMetering` and `makeGetMeter` functions
from `vatPowers` to implement metering. If `vatPowers` is not passed through
the vat `setup` function and delivered to `makeZoe()`, metering will be
disabled.