fix(spawner)!: use new SES, rewrite tests
The 'spawner' is still evolving. At present it is a subset of Zoe: it can
evaluate bundles of code in a metered/confined environment, it has
"install" (class-ish) and "spawn" (instance-ish), but it lacks Zoe's notion
of offer safety and terms and other economic nouns. It includes the
backwards-compatibility layer to allow the deploy scripts in external dapps
that still use a two-argument `install()` to continue to work, until we
rewrite them to use a single `bundle` argument.

This updates spawner to use new SES, so programs which include it must
provide a SES environment (generally by doing `import
'@agoric/install-metering-and-ses'` as their first line). It also uses the
new metering approach.