chore: remove debug tools from makeGetMeter, add usage data
I was using these during development of the within-vat metering mechanism,
and I don't need them anymore.

This also exposes a new `refill()` method, which will eventually displace use
of the `refillFacet`. This method returns the number of computrons and
allocatrons used since the last refill. Nothing pays attention to this yet,
but in the future it may be interesting to keep track of each vat's usage
over time; maybe it will help our gas model or something.